Update font conversion to use scaled size
Refactored WinFormsUtils methods to convert System.Drawing.Font to Alternet.Drawing.Font using a scale factor for font size. Added FontSizeToAlternet helper and updated relevant methods to accept and use the scale parameter for accurate font size conversion.

+++ Source/Alternet.UI.Common/Drawing/Graphics/GraphicsFactory.cs
@@ -692,7 +692,7 @@ namespace Alternet.Drawing
                 SKFontStyleWidth.Normal,
                 skiaSlant);
 
-            SKFont skiaFont = new(typeFace, (float)font.SizeInDips);
+            SKFont skiaFont = new(typeFace, font.SizeInDips);
 
             skiaFont.Subpixel = Font.DefaultSkiaFontSubpixel;
             skiaFont.Hinting = Font.DefaultSkiaFontHinting;
